+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.7.1] - 2025-12-02
+
+### Added
+
+- Formatting node address to lowercase hex
+- Formatting reg hex values to lowercase hex
+
 ## [0.7.0] - 2025-11-29
 
 ### Added